Rename container
